+++ lib/fediverse/signature/rfc9421.rb
@@ -13,7 +13,11 @@ module Fediverse
           Linzer.sign!(
             request,
             key:        linzer_key(sender),
-            components: components(request)
+            components: components(request),
+            params:     {
+              created: Time.now.to_i,
+              alg:     'rsa-v1_5-sha256',
+            }
           )
           request
         end
@@ -38,7 +42,7 @@ module Fediverse
         # Converts key to right structure for Linzer to use
         def linzer_key(sender)
           private_key = OpenSSL::PKey::RSA.new sender.private_key, Rails.application.credentials.secret_key_base
-          Linzer.new_rsa_pss_sha512_key(private_key.to_pem, sender.key_id)
+          Linzer.new_rsa_v1_5_sha256_key(private_key.to_pem, sender.key_id)
         end
 
         def components(request)